rgw: move code around
