+++ src/rgw/rgw_es_query.cc
@@ -473,32 +473,3 @@ void ESQueryCompiler::dump(Formatter *f) const {
   encode_json("query", *query_root, f);
 }
 
-
-int main(int argc, char *argv[])
-{
-  list<string> infix;
-
-  string expr;
-
-  if (argc > 1) {
-    expr = argv[1];
-  } else {
-    expr = "age >= 30";
-  }
-
-  ESQueryCompiler es_query(expr);
-  
-  bool valid = es_query.compile();
-  if (!valid) {
-    cout << "invalid query, failed generating request json" << std::endl;
-    return EINVAL;
-  }
-
-  JSONFormatter f;
-  encode_json("root", es_query, &f);
-
-  f.flush(cout);
-
-  return 0;
-}
-
